Refactor: remove dependencies.py
